Unify parser stage pipelines and DRY shared helpers.
Introduce define_stage_pipeline! for normalize/recognize/refine tables,
deduplicate grammar fields and regex/vocab/trace boilerplate, and split
large test modules out of normalize.rs and refine.rs.
